requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.3
# New dependencies for enhanced monitoring
psutil==5.9.5  # For system resource monitoring
python-json-logger==2.0.7  # Better JSON logging support

# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0 